updated in the trace query functionality

+++ src/main/java/com/zaga/entity/queryentity/trace/QueryDuration.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class QueryDuration {
-    private Integer min;
-    private Integer max;
+    private Long min;
+    private Long max;
 }
